Moved caching support stuff into the support folder
