+++ nym-api/src/circulating_supply_api/cache/mod.rs
@@ -6,7 +6,7 @@ use std::{
 use tokio::sync::RwLock;
 use tokio::time;
 
-use crate::caching_support::Cache;
+use crate::support::caching::Cache;
 
 use self::data::CirculatingSupplyCacheData;
